Fixed segfault regarding evaluation of contact_results on the input port.
The problem was that I had mistakenly created an algebraic loop in the system
that needed to be resolved. One solution was to make the system a continuous time
system by passing in a max_time_step of 0.0, and the other solution was to break
the loop between the controller and multibody plant. I opted to break the loop with
a ZOH block bc that is closer to the actual system that we are modelling. This
break was needed because the controller cannot process multibody plant outputs like
state and contact results, while simultaneously providing newly calculated inputs
to the plant.